requirements.txt
fastapi
uvicorn
groq
python-dotenv
feedparser
yfinance
pandas